separation of users for displaying del and update buttons, only an authenticated and authored user can see the buttons

+++ blog/views.py
@@ -3,6 +3,8 @@ from .models import Article
 from django.contrib.auth.decorators import login_required
 # Create your views here.
 
+from django.http import HttpResponseForbidden
+
 def index_views(request):
   return render(request, "index.html")
 
@@ -24,13 +26,13 @@ def create_article(request):
     title = request.POST["title"]
     summary = request.POST["summary"]
     content = request.POST["content"]
-    author = request.POST["author"]
+    #author = request.POST["author"]
 
     article = Article.objects.create(
       title=title,
       summary=summary,
       content=content,
-      author=author,
+      author=request.user
     )
 
     article.save()
@@ -44,17 +46,21 @@ def create_article(request):
 def update_article(request, id):
 
   articles = get_object_or_404(Article, id=id)
+
+  if request.user != articles.author:
+        return HttpResponseForbidden("Vous n'êtes pas autorisé à modifier cet article.")
+
   if request.method == "POST":
     title = request.POST['title']
     summary = request.POST['summary']
     content = request.POST['content']
-    author = request.POST['author']
+    #author = request.POST['author']
 
     Article.objects.filter(id=articles.id).update(
       title=title,
       summary=summary,
       content=content,
-      author=author
+      #author=request.user
     )
 
     return redirect('page1')
@@ -66,12 +72,16 @@ def update_article(request, id):
 @login_required(login_url="/login/")
 def delete_article(request, id):
   article = get_object_or_404(Article, id=id)
+
+  if request.user != article.author:
+    return HttpResponseForbidden("Vous n'êtes pas autorisé à supprimer cet article.")
+
   if request.method == "POST":
     article = Article.objects.filter(id=article.id)
     article.delete()
     return redirect("page1")
   
-  return render(request, "page_s/delete_article.html")
+  return render(request, "page_s/delete_article.html", {"article": article})
